refactoring INT validation

+++ app/products/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, NUMERIC
 
 from app.database import Base
 
@@ -9,5 +9,5 @@ class Products(Base):
     id = Column(Integer, primary_key=True)
     title = Column(String)
     description = Column(String)
-    price = Column(Integer)
+    price = Column(NUMERIC(10,2))
     amount = Column(Integer)
